Chore[project SETUP]
- creating .gitignore
- creating prettier.js
- creating .eslintrc.js

##[PROJECT SETUP]

+++ src/controllers/contact.js
@@ -1,11 +1,7 @@
 
-
-
-import express from 'express'
 import contact from "./../../src/models/contact"
 
-//post message
-//get all messages
+       /* ---------------------------GET ALL MESSAGES---------------------------- */
 export const getMsg =  async(req,res)=>{
     try{
         const query = await contact.find()
@@ -16,8 +12,13 @@ export const getMsg =  async(req,res)=>{
     }
 
 }
+       /* ---------------------------GET ALL MESSAGES---------------------------- */
 
-export const postMsg = async(req,res)=>{
+
+
+
+       /* ---------------------------POST MESSAGES---------------------------- */
+       export const postMsg = async(req,res)=>{
     try{
         const post = new contact({
             name:req.body.name,
@@ -31,15 +32,42 @@ export const postMsg = async(req,res)=>{
         res.send({error:"Postman not found"})
     }
 }
-//delete a message
+       /* ---------------------------POST MESSAGE---------------------------- */
+
+
+
+
+       /* ---------------------------DELETE MESSAGE---------------------------- */
+
 export const deleteContact = async(req,res)=>{
     try{
+        const contactToDelete = await contact.findOne({_id: req.params.id});
+        if (!contactToDelete) {
+            res.status(404).send({ error: 'message not found' });
+            return;
+        }
         await contact.deleteOne({_id: req.params.id})
-        res.status(204).send()
+        res.status(200).json({message:"message deleted successfully"})
     }catch{
         res.status(404)
         res.send({error:"Postman not found"})
     }
 }
+       /* ---------------------------DELETE MESSAGE ---------------------------- */
 
-module.export = postMsg
+
+
+       /* ---------------------------COUNT ALL MESSAGES---------------------------- */
+       export const countContact = async (req, res) => {
+    try {
+        const contactCount = await contact.countDocuments();
+        res.status(200).send({ message: `There are ${contactCount} contacts in the collection.` });
+    } catch (err) {
+        res.status(500).send({ error: "Error counting contacts" });
+    }
+}
+
+
+       /* ---------------------------GET ALL MESSAGES---------------------------- */
+
+module.exports = { postMsg, getMsg, deleteContact,countContact }
